refactor: move send flow components into the transfer domain
