+++ src/components/coin-row/index.js
@@ -1,5 +1,3 @@
 export { default as BottomRowText } from './BottomRowText';
 export { default as CoinRow, CoinRowHeight } from './CoinRow';
-export { default as CollectiblesSendRow } from './CollectiblesSendRow';
-export { default as SendCoinRow } from './SendCoinRow';
 export { default as FastTransactionCoinRow } from './FastTransactionCoinRow';
