+++ src/components/sheet/index.js
@@ -7,4 +7,4 @@ export { default as SheetSubtitleCycler } from './SheetSubtitleCycler';
 export { default as SheetTitle } from './SheetTitle';
 export { default as DynamicHeightSheet } from './DynamicHeightSheet';
 export { default as SlackSheet } from './SlackSheet';
-export { BuyActionButton, SendActionButton, SheetActionButton, SheetActionButtonRow, SwapActionButton } from './sheet-action-buttons';
+export { BuyActionButton, SheetActionButton, SheetActionButtonRow, SwapActionButton } from './sheet-action-buttons';
