+++ src/components/sheet/sheet-action-buttons/index.ts
@@ -1,5 +1,4 @@
 export { default as BuyActionButton } from './BuyActionButton';
-export { default as SendActionButton } from './SendActionButton';
 export { default as SheetActionButton } from './SheetActionButton';
 export { default as SheetActionButtonRow } from './SheetActionButtonRow';
 export { default as SwapActionButton } from './SwapActionButton';
