+++ src/features/transfer/components/SendAssetFormField.js
@@ -4,16 +4,15 @@ import { Platform } from 'react-native';
 import RadialGradient from 'react-native-radial-gradient';
 
 import { analytics } from '@/analytics';
+import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
+import { BubbleField } from '@/components/fields';
+import { Row, RowWithMargins } from '@/components/layout';
+import { Text } from '@/components/text';
 import styled from '@/framework/ui/styled-thing';
 import { opacity } from '@/framework/ui/utils/opacity';
 import useDimensions from '@/hooks/useDimensions';
 import * as i18n from '@/languages';
 import { useTheme } from '@/theme/ThemeContext';
-
-import ButtonPressAnimation from '../animations/ButtonPressAnimation';
-import { BubbleField } from '../fields';
-import { Row, RowWithMargins } from '../layout';
-import { Text } from '../text';
 
 const GradientBackground = styled(RadialGradient).attrs(({ colorForAsset, theme: { colors }, width }) => {
   const FieldWidth = width - 38;
