+++ src/features/transfer/components/SendAssetFormToken.js
@@ -1,6 +1,7 @@
 import React, { Fragment } from 'react';
 import { Platform } from 'react-native';
 
+import { Column } from '@/components/layout';
 import { supportedCurrencies as supportedNativeCurrencies } from '@/features/currency/supportedCurrencies';
 import styled from '@/framework/ui/styled-thing';
 import { opacity } from '@/framework/ui/utils/opacity';
@@ -9,7 +10,6 @@ import { useTheme } from '@/theme/ThemeContext';
 import { NAVIGATION_BAR_HEIGHT } from '@/utils/deviceUtils';
 import { removeLeadingZeros } from '@/utils/formatters';
 
-import { Column } from '../layout';
 import SendAssetFormField from './SendAssetFormField';
 
 const FooterContainer = styled(Column).attrs({
