+++ src/features/transfer/components/SendCoinRow.js
@@ -3,19 +3,18 @@ import { Platform, TouchableWithoutFeedback } from 'react-native';
 
 import { LinearGradient } from 'expo-linear-gradient';
 
+import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
+import RainbowCoinIcon from '@/components/coin-icon/RainbowCoinIcon';
+import CoinName from '@/components/coin-row/CoinName';
+import CoinRow from '@/components/coin-row/CoinRow';
+import { Text } from '@/components/text';
 import styled from '@/framework/ui/styled-thing';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { isL2Chain } from '@/handlers/web3';
 import useColorForAsset from '@/hooks/useColorForAsset';
 import { padding } from '@/styles';
+import { useTheme } from '@/theme/ThemeContext';
 import deviceUtils from '@/utils/deviceUtils';
-
-import { useTheme } from '../../theme/ThemeContext';
-import ButtonPressAnimation from '../animations/ButtonPressAnimation';
-import RainbowCoinIcon from '../coin-icon/RainbowCoinIcon';
-import { Text } from '../text';
-import CoinName from './CoinName';
-import CoinRow from './CoinRow';
 
 const isSmallPhone = deviceUtils.dimensions.height <= 667;
 const isTinyPhone = deviceUtils.dimensions.height <= 568;
