+++ src/features/transfer/components/SendEmptyState.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import { Platform, StyleSheet, View } from 'react-native';
 
+import { Icon } from '@/components/icons';
+import { Centered } from '@/components/layout';
 import { opacity } from '@/framework/ui/utils/opacity';
+import { sheetVerticalOffset } from '@/navigation/effects';
 import { useTheme } from '@/theme/ThemeContext';
-
-import { sheetVerticalOffset } from '../../navigation/effects';
-import { Icon } from '../icons';
-import { Centered } from '../layout';
 
 const SendEmptyState = () => {
   const { colors } = useTheme();
